+++ src/CHANGELOG.md
@@ -2,9 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.9.0] - meet project requirements
+## [1.8.0] - updated
 ## [1.7.0]
 ### added
-- fetch users using angular 19 provideHttpClient() same as products
+- fetched users using angular 19 provideHttpClient() same as products
 ## [1.6.0] - add signout
 ## [1.2.0] - 2025-03-26 🚀 (New Features & Enhancements)